requirements.txt
# Multi-Agent Chatbox 依赖包
# Python 版本要求: 3.13+
# 
# 说明：本文件中的包版本已针对 Python 3.13 进行优化
# 主要更新：
# - pydantic: 2.5.0 -> 2.10.6 (支持 Python 3.13)
# - Pillow: 10.2.0 -> 11.1.0 (支持 Python 3.13)
# - pdfplumber: 0.10.4 -> 0.11.7 (更新版本)
#
# 安装方法：
# pip install -r requirements.txt

# Web框架
fastapi==0.104.1
uvicorn==0.24.0

# AI API
openai==1.3.7

# 数据处理
python-multipart==0.0.6
pydantic==2.10.6
jinja2==3.1.2
aiofiles==23.2.1

# 文档处理
pdfplumber==0.11.7
python-docx==1.1.0
markdown==3.5.1

# 图像处理
Pillow==11.1.0
pytesseract==0.3.10

# 工具库
tenacity==8.2.3
python-dotenv==1.0.0
redis==5.0.1
slowapi==0.1.9
